lib/flaskBackend/requirements.txt
Flask
flask-cors
google-cloud-documentai>=2.0.0
requests
numpy
pandas
opencv-python
python-Levenshtein>=0.12.2
gunicorn>=20.1.0
fuzzywuzzy>=0.18.0
